requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types
astroid==3.2.4
async-timeout==5.0.1
attrs==25.1.0
bidict==0.23.1
blinker==1.8.2
certifi==2025.1.31
cffi==1.16.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coverage==7.6.1
cryptography==44.0.0
dill==0.3.9
exceptiongroup==1.2.0
flake8==7.1.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
frozenlist==1.5.0
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==1.1.1
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==2.1.5
mccabe==0.7.0
multidict==6.1.0
mypy==1.14.1
mypy-extensions==1.0.0
packaging==23.2
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.0
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.6.1
pydantic_core==2.16.2
pyflakes==3.2.0
pylint==3.2.7
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
pyqt6-tools==6.4.2.3.3
PyQt6_sip==13.8.0
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==5.0.0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
qasync==0.27.1
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
requests==2.32.3
simple-websocket==1.1.0
tomli==2.0.1
tomlkit==0.13.2
typing_extensions==4.9.0
urllib3==2.2.3
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.6
wsproto==1.2.0
yarl==1.15.2
zipp==3.20.2